backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# ML & Deep Learning
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
opencv-python==4.8.1.78
scikit-learn==1.3.2
scikit-image==0.22.0

# XAI Libraries
shap==0.43.0
lime==0.2.0.1
captum==0.7.0

# MLOps
mlflow==2.8.1
dvc==3.30.1
dvc-s3==3.0.1

# Data Processing
pandas==2.1.3
Pillow==10.1.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# Code Quality
black==23.11.0
isort==5.12.0
flake8==6.1.0
ruff==0.1.6
mypy==1.7.1

# Security & Auth
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4

# Documentation
pydantic[email]==2.5.0

# Utilities
python-dateutil==2.8.2
aiofiles==23.2.1

# Report Generation
reportlab==4.0.7
openpyxl==3.1.2
matplotlib==3.8.2
seaborn==0.13.0
